+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Hot reload option of renderer class
+
 ## [0.1.0] - 2017-09-05
 
 ### Added